+++ sdk/healthdataaiservices/azure-health-deidentification/azure/health/deidentification/_operations/_patch.py
@@ -6,33 +6,9 @@
 
 Follow our quickstart for examples: https://aka.ms/azsdk/python/dpcodegen/python/customize
 """
-from typing import Iterable, List
-from models import DocumentDetails, DeidentificationJob
-import list_jobs_internal, list_job_documents_internal
+from typing import List
 
-__all__: List[str] = ["list_jobs", "list_job_documents"]  # Add all objects you want publicly available to users at this package level
-
-def list_jobs(self) -> Iterable[DeidentificationJob]:
-    """
-    List de-identification jobs.
-
-    :return: An iterator like instance of DeidentificationJob
-    :rtype: ~azure.core.paging.ItemPaged[~azure.health.deidentification.models.DeidentificationJob]
-    :raises ~azure.core.exceptions.HttpResponseError:
-    """
-    return list_jobs_internal.list_job(self)
-
-def list_job_documents(self, job_name) -> Iterable[DocumentDetails]:
-    """
-    List processed documents within a job.
-
-    :param job_name: The name of a job. Required.
-    :type job_name: str
-    :return: An iterator like instance of DocumentDetails
-    :rtype: ~azure.core.paging.ItemPaged[~azure.health.deidentification.models.DocumentDetails]
-    :raises ~azure.core.exceptions.HttpResponseError:
-    """
-    return list_job_documents_internal.list_job_documents(self, job_name)
+__all__: List[str] = []  # Add all objects you want publicly available to users at this package level
 
 
 def patch_sdk():
